+++ server/App.ts
@@ -1,5 +1,6 @@
 import * as express from 'express';
 import splusController from './controllers/splus';
+import mensaController from './controllers/mensa';
 
 class App {
   public app: express.Application;
@@ -11,6 +12,7 @@ class App {
 
   private routes(): void {
     this.app.use(this.basePath + '/splus', splusController);
+    this.app.use(this.basePath + '/mensa', mensaController);
   }
 }
